+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Updates the pop up component to match the new designs [#152](https://github.com/PrefectHQ/miter-design/pull/152)
+- Add DataTable component [#137](https://github.com/PrefectHQ/miter-design/pull/137)
 - Add delete color option [#622](https://github.com/PrefectHQ/orion/issues/622)
 - Add Tags component [#697](https://github.com/PrefectHQ/orion/pull/697)
 - Fix bug where tabs do not update with programmatic setting of Vue ref - [#125](https://github.com/PrefectHQ/miter-design/pull/126)